vionex-audio-service/requirements.txt
faster-whisper
pydub
python-dotenv
grpcio
grpcio-tools
scipy
numpy
requests
aiohttp>=3.8.0
transformers[sentencepiece]>=4.44,<5
# torch pinned CPU-only in Dockerfile
sentencepiece
protobuf
opuslib
coqui-tts
sacremoses